Create a TaskNormalizer
The Task is the main object that is used to transfer POST and PUT
data via API HTTP requests. The normalizer uses the NameConvertor
to translate the property names relevant to the specific TaskApp.
App specific additions to the Task data are not added here to keep
the the normalizer generic. For example the Habitica 'tags' property
must be altered by the caller to be complient with their API.